requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
flask>=2.0.0
streamlit>=1.28.0
requests>=2.25.0
